Rename component
AvailableEThForm -> AvailableTokenForm to make component more general.
